add policy on community and post so that no other user can direct edit ot update using URL

+++ app/Http/Controllers/Backend/CommunityController.php
@@ -69,6 +69,7 @@ class CommunityController extends Controller
      */
     public function edit(Community $community)
     {
+        $this->authorize('update',$community);
         return Inertia::render('Communities/Edit',compact('community'));
     }
 
@@ -81,6 +82,7 @@ class CommunityController extends Controller
      */
     public function update(CommunityStoreRequest $request, Community $community)
     {
+        $this->authorize('update',$community);
         $community->update($request->validated());
 
         return to_route('communities.index')->with('message', 'Community Updated Successfully.');;
@@ -94,6 +96,7 @@ class CommunityController extends Controller
      */
     public function destroy(Community $community)
     {
+        $this->authorize('delete',$community);
         $community->delete();
         return back()->with('message', 'Community Deleted Successfully.');
     }
